Aded Consignment and stock in

+++ controllers/consignmentController.js
@@ -1,27 +1,70 @@
 const Consignment = require('../models/consignment');
 const StockIn = require('../models/stockIn');
 
-const updateQuantityByWarehouseAndCommodity = async (warehouseId, commodityId, quantity) => {
+const saveStockIn = async (warehouseId, commodityId, quantity) => {
     try {
-        const { warehouseId, commodityId, quantity } = req.body;
         const stockIn = await StockIn.findOne({ warehouseId, commodityId });
+
+        console.log('stockIn==', stockIn);
+        console.log('warehouseId, commodityId==', warehouseId, commodityId);
+
         if (!stockIn) {
-            return res.status(404).json({ status: false, message: 'StockIn not found for the given warehouse and commodity' });
+            const newStockIn = new StockIn({ warehouseId, commodityId, quantity });
+            await newStockIn.save();
+        } else {
+            stockIn.quantity += quantity;
+            await stockIn.save();
         }
-        stockIn.quantity = quantity;
-        await stockIn.save();
-        res.json({ status: true, message: 'Quantity updated successfully', data: stockIn });
+
     } catch (error) {
-        res.status(500).json({ status: false, message: 'Failed to update quantity', error: error.message });
+        console.log('message: ', 'Failed to update quantity', 'error:', error.message);
     }
-};
+}
+
+const deleteStockIn = async (warehouseId, commodityId, quantity) => {
+    try {
+        const stockIn = await StockIn.findOne({ warehouseId, commodityId });
+
+        if (stockIn) {
+            stockIn.quantity = stockIn.quantity - quantity;
+
+            if (stockIn.quantity < 0) {
+                stockIn.quantity = 0;
+            }
+            await stockIn.save();
+        }
+
+    } catch (error) {
+        console.log('message: ', 'Failed to update quantity', 'error:', error.message);
+    }
+}
+
+// const updateStockIn = async (warehouseId, commodityId, quantity) => {
+//     try {
+//         const stockIn = await StockIn.findOne({ warehouseId, commodityId });
+
+//         console.log('stockIn==', stockIn);
+//         console.log('warehouseId, commodityId==', warehouseId, commodityId);
+
+//         if (!stockIn) {
+//             const newStockIn = new StockIn({ warehouseId, commodityId, quantity });
+//             await newStockIn.save();
+//         } else {
+//             stockIn.quantity += quantity;
+//             await stockIn.save();
+//         }
+
+//     } catch (error) {
+//         console.log('message: ', 'Failed to update quantity', 'error:', error.message);
+//     }
+// };
 
 const createConsignment = async (req, res, next) => {
     try {
         const { farmerId, transporterId, warehouseId, commodityId, quantity, rate, amount } = req.body;
         const newConsignment = new Consignment({ farmerId, transporterId, warehouseId, commodityId, quantity, rate, amount });
         await newConsignment.save();
-        await updateQuantityByWarehouseAndCommodity(warehouseId, commodityId, quantity);
+        await saveStockIn(warehouseId, commodityId, quantity);
         res.status(201).json({ status: true, message: 'Consignment created successfully', data: newConsignment });
     } catch (error) {
         res.status(400).json({ status: false, message: 'Failed to create consignment', error: error.message });
@@ -56,7 +99,13 @@ const updateConsignment = async (req, res, next) => {
         if (!consignment) {
             return res.status(404).json({ status: false, message: 'Consignment not found' });
         }
-        await updateQuantityByWarehouseAndCommodity(warehouseId, commodityId, quantity);
+
+        // if (Number(consignment.quantity) > Number(quantity)) {
+        //     let newQuantity = Number(consignment.quantity) - Number(quantity)
+        // } else {
+        //     let newQuantity = Number(consignment.quantity) - Number(quantity)
+        // }
+        // await updateStockIn(warehouseId, commodityId, quantity);
         res.json({ status: true, message: 'Consignment updated successfully', data: consignment });
     } catch (error) {
         res.status(500).json({ status: false, message: 'Failed to update consignment', error: error.message });
@@ -65,12 +114,18 @@ const updateConsignment = async (req, res, next) => {
 
 const deleteConsignment = async (req, res, next) => {
     try {
+
+        const consignment = await Consignment.findById(req.params.id);
+        if (!consignment) {
+            return res.status(404).json({ status: false, message: 'Consignment not found' });
+        }
+
         const deletedConsignment = await Consignment.findByIdAndDelete(req.params.id);
         if (!deletedConsignment) {
             return res.status(404).json({ status: false, message: 'Consignment not found' });
         }
 
-        await updateQuantityByWarehouseAndCommodity(warehouseId, commodityId, 0);
+        await deleteStockIn(consignment.warehouseId, consignment.commodityId, consignment.quantity)
 
         res.json({ status: true, message: 'Consignment deleted successfully' });
     } catch (error) {
